fix: duplication of external layers
